% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashlookup-package.R
\docType{package}
\name{hashlookup}
\alias{hashlookup}
\alias{hashlookup-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/hashlookup}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/hashlookup/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
