% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl-info.R
\name{hl_info}
\alias{hl_info}
\alias{print.hl_info}
\title{Get information about the hash lookup database}
\usage{
hl_info()

\method{print}{hl_info}(x, ...)
}
\arguments{
\item{x}{hashlookup info object}

\item{...}{unused}
}
\value{
list classes \code{hl_info}
}
\description{
Get information about the hash lookup database
}
\examples{
hl_info()
}
\references{
\href{https://circl.lu/services/hashlookup/}{CIRCL Hash Lookup}
}
