% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{hl_query}
\alias{hl_query}
\title{Get information about a hash}
\usage{
hl_query(hash, type = c("md5", "sha1", "sha256"))
}
\arguments{
\item{hash}{(chr) length can be > 1 for bulk queries}

\item{type}{(chr) hash type. All passed-in hashes must be of the same type. Default is \code{md5}.}
}
\value{
data frame for \code{md5} and \code{sha1}; list for \code{sha256}
}
\description{
Get information about a hash
}
\note{
SHA-256 bulk lookups are not supported by the CIRCL hash lookup API
}
\examples{
# individual

hl_query("8ED4B4ED952526D89899E723F3488DE4", "md5")
hl_query("FFFFFDAC1B1B4C513896C805C2C698D9688BE69F", "sha1")
hl_query("301c9ec7a9aadee4d745e8fd4fa659dafbbcc6b75b9ff491d14cbbdd840814e9", "sha256")

# bulk

hl_query(
  hash = c(
    "6E2F8616A01725DCB37BED0A2495AEB2",
    "8ED4B4ED952526D89899E723F3488DE4",
    "344428FA4BA313712E4CA9B16D089AC4"
  ),
  type = "md5"
)

}
\references{
\href{https://circl.lu/services/hashlookup/}{CIRCL Hash Lookup}
}
