% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashlookup-package.R
\docType{package}
\name{hashlookup}
\alias{hashlookup}
\alias{hashlookup-package}
\title{Query and Orchestrate the CIRCL Hash Lookup Service}
\description{
CIRCL hash lookup (\url{https://circl.lu/services/hashlookup/}) is a public API to lookup
hash values against known database of files. National Software Reference Library (NSRL)
Reference Data Set (RDS) database is included and many others are
also included. Tools are provided to query and orchestrate the CIRCL hash lookup service.
}
\references{
\href{https://circl.lu/services/hashlookup/}{CIRCL Hash Lookup}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrbrmstr/hashlookup}
  \item Report bugs at \url{https://github.com/hrbrmstr/hashlookup/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
